refactor: changed Embox to its own PrettyPrint class
